msm8974: Force regeneration of sensor registry on every boot
Currently, our proprietary sensors library will only create the sensor registry when it is not present, otherwise it'll just use the registry that already exists. Due to this, changes to the default sensor registry values would not be propagated to the sensor registry unless the existing registry were deleted or the end user performed a factory reset.

Force regeneration of the sensor registry on every boot to ensure that it is always up-to-date.